requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch
torchvision
torchaudio

nerfstudio

numpy>=1.26
opencv-python>=4.8
pyyaml>=6.0
tqdm>=4.66
